% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{AddOrderDealsRequest}
\alias{AddOrderDealsRequest}
\title{AddOrderDealsRequest Object}
\usage{
AddOrderDealsRequest(deals = NULL, proposalRevisionNumber = NULL,
  updateAction = NULL)
}
\arguments{
\item{deals}{The list of deals to add}

\item{proposalRevisionNumber}{The last known proposal revision number}

\item{updateAction}{Indicates an optional action to take on the proposal}
}
\value{
AddOrderDealsRequest object
}
\description{
AddOrderDealsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AddOrderDealsRequest functions: \code{\link{marketplacedeals.insert}}
}

