% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{AddOrderDealsResponse}
\alias{AddOrderDealsResponse}
\title{AddOrderDealsResponse Object}
\usage{
AddOrderDealsResponse(deals = NULL, proposalRevisionNumber = NULL)
}
\arguments{
\item{deals}{List of deals added (in the same proposal as passed in the request)}

\item{proposalRevisionNumber}{The updated revision number for the proposal}
}
\value{
AddOrderDealsResponse object
}
\description{
AddOrderDealsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
