% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{AddOrderNotesRequest}
\alias{AddOrderNotesRequest}
\title{AddOrderNotesRequest Object}
\usage{
AddOrderNotesRequest(notes = NULL)
}
\arguments{
\item{notes}{The list of notes to add}
}
\value{
AddOrderNotesRequest object
}
\description{
AddOrderNotesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AddOrderNotesRequest functions: \code{\link{marketplacenotes.insert}}
}

