% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{AddOrderNotesResponse}
\alias{AddOrderNotesResponse}
\title{AddOrderNotesResponse Object}
\usage{
AddOrderNotesResponse(notes = NULL)
}
\arguments{
\item{notes}{No description}
}
\value{
AddOrderNotesResponse object
}
\description{
AddOrderNotesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

