% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Buyer}
\alias{Buyer}
\title{Buyer Object}
\usage{
Buyer(accountId = NULL)
}
\arguments{
\item{accountId}{Adx account id of the buyer}
}
\value{
Buyer object
}
\description{
Buyer Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

