% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{ContactInformation}
\alias{ContactInformation}
\title{ContactInformation Object}
\usage{
ContactInformation(email = NULL, name = NULL)
}
\arguments{
\item{email}{Email address of the contact}

\item{name}{The name of the contact}
}
\value{
ContactInformation object
}
\description{
ContactInformation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

