% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{CreateOrdersRequest}
\alias{CreateOrdersRequest}
\title{CreateOrdersRequest Object}
\usage{
CreateOrdersRequest(proposals = NULL, webPropertyCode = NULL)
}
\arguments{
\item{proposals}{The list of proposals to create}

\item{webPropertyCode}{Web property id of the seller creating these orders}
}
\value{
CreateOrdersRequest object
}
\description{
CreateOrdersRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other CreateOrdersRequest functions: \code{\link{proposals.insert}}
}
