% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{CreateOrdersResponse}
\alias{CreateOrdersResponse}
\title{CreateOrdersResponse Object}
\usage{
CreateOrdersResponse(proposals = NULL)
}
\arguments{
\item{proposals}{The list of proposals successfully created}
}
\value{
CreateOrdersResponse object
}
\description{
CreateOrdersResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
