% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Creative.filteringReasons.reasons}
\alias{Creative.filteringReasons.reasons}
\title{Creative.filteringReasons.reasons Object}
\usage{
Creative.filteringReasons.reasons()
}
\value{
Creative.filteringReasons.reasons object
}
\description{
Creative.filteringReasons.reasons Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Creative functions: \code{\link{Creative.corrections.contexts}},
  \code{\link{Creative.corrections}},
  \code{\link{Creative.filteringReasons}},
  \code{\link{Creative.nativeAd.appIcon}},
  \code{\link{Creative.nativeAd.image}},
  \code{\link{Creative.nativeAd.logo}},
  \code{\link{Creative.nativeAd}},
  \code{\link{Creative.servingRestrictions.contexts}},
  \code{\link{Creative.servingRestrictions.disapprovalReasons}},
  \code{\link{Creative.servingRestrictions}},
  \code{\link{Creative}}, \code{\link{creatives.insert}}
}

