% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Creative.nativeAd.image}
\alias{Creative.nativeAd.image}
\title{Creative.nativeAd.image Object}
\usage{
Creative.nativeAd.image(height = NULL, url = NULL, width = NULL)
}
\arguments{
\item{height}{No description}

\item{url}{No description}

\item{width}{No description}
}
\value{
Creative.nativeAd.image object
}
\description{
Creative.nativeAd.image Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A large image.
}
\seealso{
Other Creative functions: \code{\link{Creative.corrections.contexts}},
  \code{\link{Creative.corrections}},
  \code{\link{Creative.filteringReasons.reasons}},
  \code{\link{Creative.filteringReasons}},
  \code{\link{Creative.nativeAd.appIcon}},
  \code{\link{Creative.nativeAd.logo}},
  \code{\link{Creative.nativeAd}},
  \code{\link{Creative.servingRestrictions.contexts}},
  \code{\link{Creative.servingRestrictions.disapprovalReasons}},
  \code{\link{Creative.servingRestrictions}},
  \code{\link{Creative}}, \code{\link{creatives.insert}}
}
