% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Creative.nativeAd}
\alias{Creative.nativeAd}
\title{Creative.nativeAd Object}
\usage{
Creative.nativeAd(Creative.nativeAd.appIcon = NULL,
  Creative.nativeAd.image = NULL, Creative.nativeAd.logo = NULL,
  advertiser = NULL, appIcon = NULL, body = NULL, callToAction = NULL,
  clickLinkUrl = NULL, clickTrackingUrl = NULL, headline = NULL,
  image = NULL, impressionTrackingUrl = NULL, logo = NULL, price = NULL,
  starRating = NULL, store = NULL, videoURL = NULL)
}
\arguments{
\item{Creative.nativeAd.appIcon}{The \link{Creative.nativeAd.appIcon} object or list of objects}

\item{Creative.nativeAd.image}{The \link{Creative.nativeAd.image} object or list of objects}

\item{Creative.nativeAd.logo}{The \link{Creative.nativeAd.logo} object or list of objects}

\item{advertiser}{No description}

\item{appIcon}{The app icon, for app download ads}

\item{body}{A long description of the ad}

\item{callToAction}{A label for the button that the user is supposed to click}

\item{clickLinkUrl}{The URL that the browser/SDK will load when the user clicks the ad}

\item{clickTrackingUrl}{The URL to use for click tracking}

\item{headline}{A short title for the ad}

\item{image}{A large image}

\item{impressionTrackingUrl}{The URLs are called when the impression is rendered}

\item{logo}{A smaller image, for the advertiser logo}

\item{price}{The price of the promoted app including the currency info}

\item{starRating}{The app rating in the app store}

\item{store}{The URL to the app store to purchase/download the promoted app}

\item{videoURL}{The URL of the XML VAST for a native ad}
}
\value{
Creative.nativeAd object
}
\description{
Creative.nativeAd Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
If nativeAd is set, HTMLSnippet and the videoURL outside of nativeAd should not be set. (The videoURL inside nativeAd can be set.)
}
\seealso{
Other Creative functions: \code{\link{Creative.corrections.contexts}},
  \code{\link{Creative.corrections}},
  \code{\link{Creative.filteringReasons.reasons}},
  \code{\link{Creative.filteringReasons}},
  \code{\link{Creative.nativeAd.appIcon}},
  \code{\link{Creative.nativeAd.image}},
  \code{\link{Creative.nativeAd.logo}},
  \code{\link{Creative.servingRestrictions.contexts}},
  \code{\link{Creative.servingRestrictions.disapprovalReasons}},
  \code{\link{Creative.servingRestrictions}},
  \code{\link{Creative}}, \code{\link{creatives.insert}}
}
