% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Creative}
\alias{Creative}
\title{Creative Object}
\usage{
Creative(Creative.corrections = NULL, Creative.corrections.contexts = NULL,
  Creative.filteringReasons = NULL,
  Creative.filteringReasons.reasons = NULL, Creative.nativeAd = NULL,
  Creative.nativeAd.appIcon = NULL, Creative.nativeAd.image = NULL,
  Creative.nativeAd.logo = NULL, Creative.servingRestrictions = NULL,
  Creative.servingRestrictions.contexts = NULL,
  Creative.servingRestrictions.disapprovalReasons = NULL,
  HTMLSnippet = NULL, accountId = NULL, adChoicesDestinationUrl = NULL,
  advertiserId = NULL, advertiserName = NULL, agencyId = NULL,
  apiUploadTimestamp = NULL, attribute = NULL, buyerCreativeId = NULL,
  clickThroughUrl = NULL, corrections = NULL, dealsStatus = NULL,
  detectedDomains = NULL, filteringReasons = NULL, height = NULL,
  impressionTrackingUrl = NULL, languages = NULL, nativeAd = NULL,
  openAuctionStatus = NULL, productCategories = NULL,
  restrictedCategories = NULL, sensitiveCategories = NULL,
  servingRestrictions = NULL, vendorType = NULL, version = NULL,
  videoURL = NULL, width = NULL)
}
\arguments{
\item{Creative.corrections}{The \link{Creative.corrections} object or list of objects}

\item{Creative.corrections.contexts}{The \link{Creative.corrections.contexts} object or list of objects}

\item{Creative.filteringReasons}{The \link{Creative.filteringReasons} object or list of objects}

\item{Creative.filteringReasons.reasons}{The \link{Creative.filteringReasons.reasons} object or list of objects}

\item{Creative.nativeAd}{The \link{Creative.nativeAd} object or list of objects}

\item{Creative.nativeAd.appIcon}{The \link{Creative.nativeAd.appIcon} object or list of objects}

\item{Creative.nativeAd.image}{The \link{Creative.nativeAd.image} object or list of objects}

\item{Creative.nativeAd.logo}{The \link{Creative.nativeAd.logo} object or list of objects}

\item{Creative.servingRestrictions}{The \link{Creative.servingRestrictions} object or list of objects}

\item{Creative.servingRestrictions.contexts}{The \link{Creative.servingRestrictions.contexts} object or list of objects}

\item{Creative.servingRestrictions.disapprovalReasons}{The \link{Creative.servingRestrictions.disapprovalReasons} object or list of objects}

\item{HTMLSnippet}{The HTML snippet that displays the ad when inserted in the web page}

\item{accountId}{Account id}

\item{adChoicesDestinationUrl}{The link to the Ad Preferences page}

\item{advertiserId}{Detected advertiser id, if any}

\item{advertiserName}{The name of the company being advertised in the creative}

\item{agencyId}{The agency id for this creative}

\item{apiUploadTimestamp}{The last upload timestamp of this creative if it was uploaded via API}

\item{attribute}{List of buyer selectable attributes for the ads that may be shown from this snippet}

\item{buyerCreativeId}{A buyer-specific id identifying the creative in this ad}

\item{clickThroughUrl}{The set of destination urls for the snippet}

\item{corrections}{Shows any corrections that were applied to this creative}

\item{dealsStatus}{Top-level deals status}

\item{detectedDomains}{Detected domains for this creative}

\item{filteringReasons}{The filtering reasons for the creative}

\item{height}{Ad height}

\item{impressionTrackingUrl}{The set of urls to be called to record an impression}

\item{languages}{Detected languages for this creative}

\item{nativeAd}{If nativeAd is set, HTMLSnippet and the videoURL outside of nativeAd should not be set}

\item{openAuctionStatus}{Top-level open auction status}

\item{productCategories}{Detected product categories, if any}

\item{restrictedCategories}{All restricted categories for the ads that may be shown from this snippet}

\item{sensitiveCategories}{Detected sensitive categories, if any}

\item{servingRestrictions}{The granular status of this ad in specific contexts}

\item{vendorType}{List of vendor types for the ads that may be shown from this snippet}

\item{version}{The version for this creative}

\item{videoURL}{The URL to fetch a video ad}

\item{width}{Ad width}
}
\value{
Creative object
}
\description{
Creative Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A creative and its classification data.
}
\seealso{
Other Creative functions: \code{\link{Creative.corrections.contexts}},
  \code{\link{Creative.corrections}},
  \code{\link{Creative.filteringReasons.reasons}},
  \code{\link{Creative.filteringReasons}},
  \code{\link{Creative.nativeAd.appIcon}},
  \code{\link{Creative.nativeAd.image}},
  \code{\link{Creative.nativeAd.logo}},
  \code{\link{Creative.nativeAd}},
  \code{\link{Creative.servingRestrictions.contexts}},
  \code{\link{Creative.servingRestrictions.disapprovalReasons}},
  \code{\link{Creative.servingRestrictions}},
  \code{\link{creatives.insert}}
}
