% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Creative.servingRestrictions.disapprovalReasons}
\alias{Creative.servingRestrictions.disapprovalReasons}
\title{Creative.servingRestrictions.disapprovalReasons Object}
\usage{
Creative.servingRestrictions.disapprovalReasons()
}
\value{
Creative.servingRestrictions.disapprovalReasons object
}
\description{
Creative.servingRestrictions.disapprovalReasons Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Creative functions: \code{\link{Creative.corrections.contexts}},
  \code{\link{Creative.corrections}},
  \code{\link{Creative.filteringReasons.reasons}},
  \code{\link{Creative.filteringReasons}},
  \code{\link{Creative.nativeAd.appIcon}},
  \code{\link{Creative.nativeAd.image}},
  \code{\link{Creative.nativeAd.logo}},
  \code{\link{Creative.nativeAd}},
  \code{\link{Creative.servingRestrictions.contexts}},
  \code{\link{Creative.servingRestrictions}},
  \code{\link{Creative}}, \code{\link{creatives.insert}}
}
