% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Creative.servingRestrictions}
\alias{Creative.servingRestrictions}
\title{Creative.servingRestrictions Object}
\usage{
Creative.servingRestrictions(Creative.servingRestrictions.contexts = NULL,
  Creative.servingRestrictions.disapprovalReasons = NULL)
}
\arguments{
\item{Creative.servingRestrictions.contexts}{The \link{Creative.servingRestrictions.contexts} object or list of objects}

\item{Creative.servingRestrictions.disapprovalReasons}{The \link{Creative.servingRestrictions.disapprovalReasons} object or list of objects}
}
\value{
Creative.servingRestrictions object
}
\description{
Creative.servingRestrictions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Creative functions: \code{\link{Creative.corrections.contexts}},
  \code{\link{Creative.corrections}},
  \code{\link{Creative.filteringReasons.reasons}},
  \code{\link{Creative.filteringReasons}},
  \code{\link{Creative.nativeAd.appIcon}},
  \code{\link{Creative.nativeAd.image}},
  \code{\link{Creative.nativeAd.logo}},
  \code{\link{Creative.nativeAd}},
  \code{\link{Creative.servingRestrictions.contexts}},
  \code{\link{Creative.servingRestrictions.disapprovalReasons}},
  \code{\link{Creative}}, \code{\link{creatives.insert}}
}
