% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{CreativeDealIds.dealStatuses}
\alias{CreativeDealIds.dealStatuses}
\title{CreativeDealIds.dealStatuses Object}
\usage{
CreativeDealIds.dealStatuses()
}
\value{
CreativeDealIds.dealStatuses object
}
\description{
CreativeDealIds.dealStatuses Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other CreativeDealIds functions: \code{\link{CreativeDealIds}}
}
