% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{CreativeDealIds}
\alias{CreativeDealIds}
\title{CreativeDealIds Object}
\usage{
CreativeDealIds(CreativeDealIds.dealStatuses = NULL, dealStatuses = NULL)
}
\arguments{
\item{CreativeDealIds.dealStatuses}{The \link{CreativeDealIds.dealStatuses} object or list of objects}

\item{dealStatuses}{A list of external deal ids and ARC approval status}
}
\value{
CreativeDealIds object
}
\description{
CreativeDealIds Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The external deal ids associated with a creative.
}
\seealso{
Other CreativeDealIds functions: \code{\link{CreativeDealIds.dealStatuses}}
}
