% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_functions.R
\name{creatives.removeDeal}
\alias{creatives.removeDeal}
\title{Remove a deal id associated with the creative.}
\usage{
creatives.removeDeal(accountId, buyerCreativeId, dealId)
}
\arguments{
\item{accountId}{The id for the account that will serve this creative}

\item{buyerCreativeId}{The buyer-specific id for this creative}

\item{dealId}{The id of the deal id to disassociate with this creative}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adexchange.buyer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/ad-exchange/buyer-rest}{Google Documentation}
}

