% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{DealServingMetadata}
\alias{DealServingMetadata}
\title{DealServingMetadata Object}
\usage{
DealServingMetadata(alcoholAdsAllowed = NULL, dealPauseStatus = NULL)
}
\arguments{
\item{alcoholAdsAllowed}{True if alcohol ads are allowed for this deal (read-only)}

\item{dealPauseStatus}{Tracks which parties (if any) have paused a deal}
}
\value{
DealServingMetadata object
}
\description{
DealServingMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
