% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{DealServingMetadataDealPauseStatus}
\alias{DealServingMetadataDealPauseStatus}
\title{DealServingMetadataDealPauseStatus Object}
\usage{
DealServingMetadataDealPauseStatus(buyerPauseReason = NULL,
  firstPausedBy = NULL, hasBuyerPaused = NULL, hasSellerPaused = NULL,
  sellerPauseReason = NULL)
}
\arguments{
\item{buyerPauseReason}{No description}

\item{firstPausedBy}{If the deal is paused, records which party paused the deal first}

\item{hasBuyerPaused}{No description}

\item{hasSellerPaused}{No description}

\item{sellerPauseReason}{No description}
}
\value{
DealServingMetadataDealPauseStatus object
}
\description{
DealServingMetadataDealPauseStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Tracks which parties (if any) have paused a deal. The deal is considered paused if has_buyer_paused || has_seller_paused. Each of the has_buyer_paused or the has_seller_paused bits can be set independently.
}
