% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{DealTerms}
\alias{DealTerms}
\title{DealTerms Object}
\usage{
DealTerms(brandingType = NULL, crossListedExternalDealIdType = NULL,
  description = NULL, estimatedGrossSpend = NULL,
  estimatedImpressionsPerDay = NULL, guaranteedFixedPriceTerms = NULL,
  nonGuaranteedAuctionTerms = NULL, nonGuaranteedFixedPriceTerms = NULL,
  rubiconNonGuaranteedTerms = NULL, sellerTimeZone = NULL)
}
\arguments{
\item{brandingType}{Visibilty of the URL in bid requests}

\item{crossListedExternalDealIdType}{Indicates that this ExternalDealId exists under at least two different AdxInventoryDeals}

\item{description}{Description for the proposed terms of the deal}

\item{estimatedGrossSpend}{Non-binding estimate of the estimated gross spend for this deal Can be set by buyer or seller}

\item{estimatedImpressionsPerDay}{Non-binding estimate of the impressions served per day Can be set by buyer or seller}

\item{guaranteedFixedPriceTerms}{The terms for guaranteed fixed price deals}

\item{nonGuaranteedAuctionTerms}{The terms for non-guaranteed auction deals}

\item{nonGuaranteedFixedPriceTerms}{The terms for non-guaranteed fixed price deals}

\item{rubiconNonGuaranteedTerms}{The terms for rubicon non-guaranteed deals}

\item{sellerTimeZone}{For deals with Cost Per Day billing, defines the timezone used to mark the boundaries of a day (buyer-readonly)}
}
\value{
DealTerms object
}
\description{
DealTerms Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

