% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{DealTermsGuaranteedFixedPriceTerms}
\alias{DealTermsGuaranteedFixedPriceTerms}
\title{DealTermsGuaranteedFixedPriceTerms Object}
\usage{
DealTermsGuaranteedFixedPriceTerms(billingInfo = NULL, fixedPrices = NULL,
  guaranteedImpressions = NULL, guaranteedLooks = NULL,
  minimumDailyLooks = NULL)
}
\arguments{
\item{billingInfo}{External billing info for this Deal}

\item{fixedPrices}{Fixed price for the specified buyer}

\item{guaranteedImpressions}{Guaranteed impressions as a percentage}

\item{guaranteedLooks}{Count of guaranteed looks}

\item{minimumDailyLooks}{Count of minimum daily looks for a CPD deal}
}
\value{
DealTermsGuaranteedFixedPriceTerms object
}
\description{
DealTermsGuaranteedFixedPriceTerms Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
