% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{DealTermsGuaranteedFixedPriceTermsBillingInfo}
\alias{DealTermsGuaranteedFixedPriceTermsBillingInfo}
\title{DealTermsGuaranteedFixedPriceTermsBillingInfo Object}
\usage{

  DealTermsGuaranteedFixedPriceTermsBillingInfo(currencyConversionTimeMs = NULL,
  dfpLineItemId = NULL, originalContractedQuantity = NULL, price = NULL)
}
\arguments{
\item{currencyConversionTimeMs}{The timestamp (in ms since epoch) when the original reservation price for the deal was first converted to DFP currency}

\item{dfpLineItemId}{The DFP line item id associated with this deal}

\item{originalContractedQuantity}{The original contracted quantity (# impressions) for this deal}

\item{price}{The original reservation price for the deal, if the currency code is different from the one used in negotiation}
}
\value{
DealTermsGuaranteedFixedPriceTermsBillingInfo object
}
\description{
DealTermsGuaranteedFixedPriceTermsBillingInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
