% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{DealTermsNonGuaranteedAuctionTerms}
\alias{DealTermsNonGuaranteedAuctionTerms}
\title{DealTermsNonGuaranteedAuctionTerms Object}
\usage{
DealTermsNonGuaranteedAuctionTerms(autoOptimizePrivateAuction = NULL,
  reservePricePerBuyers = NULL)
}
\arguments{
\item{autoOptimizePrivateAuction}{True if open auction buyers are allowed to compete with invited buyers in this private auction (buyer-readonly)}

\item{reservePricePerBuyers}{Reserve price for the specified buyer}
}
\value{
DealTermsNonGuaranteedAuctionTerms object
}
\description{
DealTermsNonGuaranteedAuctionTerms Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
