% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{DealTermsNonGuaranteedFixedPriceTerms}
\alias{DealTermsNonGuaranteedFixedPriceTerms}
\title{DealTermsNonGuaranteedFixedPriceTerms Object}
\usage{
DealTermsNonGuaranteedFixedPriceTerms(fixedPrices = NULL)
}
\arguments{
\item{fixedPrices}{Fixed price for the specified buyer}
}
\value{
DealTermsNonGuaranteedFixedPriceTerms object
}
\description{
DealTermsNonGuaranteedFixedPriceTerms Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
