% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{DealTermsRubiconNonGuaranteedTerms}
\alias{DealTermsRubiconNonGuaranteedTerms}
\title{DealTermsRubiconNonGuaranteedTerms Object}
\usage{
DealTermsRubiconNonGuaranteedTerms(priorityPrice = NULL,
  standardPrice = NULL)
}
\arguments{
\item{priorityPrice}{Optional price for Rubicon priority access in the auction}

\item{standardPrice}{Optional price for Rubicon standard access in the auction}
}
\value{
DealTermsRubiconNonGuaranteedTerms object
}
\description{
DealTermsRubiconNonGuaranteedTerms Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

