% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{DeleteOrderDealsRequest}
\alias{DeleteOrderDealsRequest}
\title{DeleteOrderDealsRequest Object}
\usage{
DeleteOrderDealsRequest(dealIds = NULL, proposalRevisionNumber = NULL,
  updateAction = NULL)
}
\arguments{
\item{dealIds}{List of deals to delete for a given proposal}

\item{proposalRevisionNumber}{The last known proposal revision number}

\item{updateAction}{Indicates an optional action to take on the proposal}
}
\value{
DeleteOrderDealsRequest object
}
\description{
DeleteOrderDealsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other DeleteOrderDealsRequest functions: \code{\link{marketplacedeals.delete}}
}

