% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{DeleteOrderDealsResponse}
\alias{DeleteOrderDealsResponse}
\title{DeleteOrderDealsResponse Object}
\usage{
DeleteOrderDealsResponse(deals = NULL, proposalRevisionNumber = NULL)
}
\arguments{
\item{deals}{List of deals deleted (in the same proposal as passed in the request)}

\item{proposalRevisionNumber}{The updated revision number for the proposal}
}
\value{
DeleteOrderDealsResponse object
}
\description{
DeleteOrderDealsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

