% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{DeliveryControl}
\alias{DeliveryControl}
\title{DeliveryControl Object}
\usage{
DeliveryControl(creativeBlockingLevel = NULL, deliveryRateType = NULL,
  frequencyCaps = NULL)
}
\arguments{
\item{creativeBlockingLevel}{No description}

\item{deliveryRateType}{No description}

\item{frequencyCaps}{No description}
}
\value{
DeliveryControl object
}
\description{
DeliveryControl Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
