% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{DeliveryControlFrequencyCap}
\alias{DeliveryControlFrequencyCap}
\title{DeliveryControlFrequencyCap Object}
\usage{
DeliveryControlFrequencyCap(maxImpressions = NULL, numTimeUnits = NULL,
  timeUnitType = NULL)
}
\arguments{
\item{maxImpressions}{No description}

\item{numTimeUnits}{No description}

\item{timeUnitType}{No description}
}
\value{
DeliveryControlFrequencyCap object
}
\description{
DeliveryControlFrequencyCap Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
