% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Dimension}
\alias{Dimension}
\title{Dimension Object}
\usage{
Dimension(dimensionType = NULL, dimensionValues = NULL)
}
\arguments{
\item{dimensionType}{No description}

\item{dimensionValues}{No description}
}
\value{
Dimension object
}
\description{
Dimension Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This message carries publisher provided breakdown. E.g. {dimension_type: 'COUNTRY', [{dimension_value: {id: 1, name: 'US'}}, {dimension_value: {id: 2, name: 'UK'}}]}
}

