% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{DimensionDimensionValue}
\alias{DimensionDimensionValue}
\title{DimensionDimensionValue Object}
\usage{
DimensionDimensionValue(id = NULL, name = NULL, percentage = NULL)
}
\arguments{
\item{id}{Id of the dimension}

\item{name}{Name of the dimension mainly for debugging purposes, except for the case of CREATIVE_SIZE}

\item{percentage}{Percent of total impressions for a dimension type}
}
\value{
DimensionDimensionValue object
}
\description{
DimensionDimensionValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Value of the dimension.
}

