% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{EditAllOrderDealsRequest}
\alias{EditAllOrderDealsRequest}
\title{EditAllOrderDealsRequest Object}
\usage{
EditAllOrderDealsRequest(deals = NULL, proposal = NULL,
  proposalRevisionNumber = NULL, updateAction = NULL)
}
\arguments{
\item{deals}{List of deals to edit}

\item{proposal}{If specified, also updates the proposal in the batch transaction}

\item{proposalRevisionNumber}{The last known revision number for the proposal}

\item{updateAction}{Indicates an optional action to take on the proposal}
}
\value{
EditAllOrderDealsRequest object
}
\description{
EditAllOrderDealsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other EditAllOrderDealsRequest functions: \code{\link{marketplacedeals.update}}
}

