% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{EditAllOrderDealsResponse}
\alias{EditAllOrderDealsResponse}
\title{EditAllOrderDealsResponse Object}
\usage{
EditAllOrderDealsResponse(deals = NULL, orderRevisionNumber = NULL)
}
\arguments{
\item{deals}{List of all deals in the proposal after edit}

\item{orderRevisionNumber}{The latest revision number after the update has been applied}
}
\value{
EditAllOrderDealsResponse object
}
\description{
EditAllOrderDealsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
