% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{GetOffersResponse}
\alias{GetOffersResponse}
\title{GetOffersResponse Object}
\usage{
GetOffersResponse(products = NULL)
}
\arguments{
\item{products}{The returned list of products}
}
\value{
GetOffersResponse object
}
\description{
GetOffersResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
