% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{GetOrderDealsResponse}
\alias{GetOrderDealsResponse}
\title{GetOrderDealsResponse Object}
\usage{
GetOrderDealsResponse(deals = NULL)
}
\arguments{
\item{deals}{List of deals for the proposal}
}
\value{
GetOrderDealsResponse object
}
\description{
GetOrderDealsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
