% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{GetOrderNotesResponse}
\alias{GetOrderNotesResponse}
\title{GetOrderNotesResponse Object}
\usage{
GetOrderNotesResponse(notes = NULL)
}
\arguments{
\item{notes}{The list of matching notes}
}
\value{
GetOrderNotesResponse object
}
\description{
GetOrderNotesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

