% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{GetOrdersResponse}
\alias{GetOrdersResponse}
\title{GetOrdersResponse Object}
\usage{
GetOrdersResponse(proposals = NULL)
}
\arguments{
\item{proposals}{The list of matching proposals}
}
\value{
GetOrdersResponse object
}
\description{
GetOrdersResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
