% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{GetPublisherProfilesByAccountIdResponse}
\alias{GetPublisherProfilesByAccountIdResponse}
\title{GetPublisherProfilesByAccountIdResponse Object}
\usage{
GetPublisherProfilesByAccountIdResponse(profiles = NULL)
}
\arguments{
\item{profiles}{Profiles for the requested publisher}
}
\value{
GetPublisherProfilesByAccountIdResponse object
}
\description{
GetPublisherProfilesByAccountIdResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
