% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{MarketplaceDeal}
\alias{MarketplaceDeal}
\title{MarketplaceDeal Object}
\usage{
MarketplaceDeal(buyerPrivateData = NULL, creationTimeMs = NULL,
  creativePreApprovalPolicy = NULL, creativeSafeFrameCompatibility = NULL,
  dealId = NULL, dealServingMetadata = NULL, deliveryControl = NULL,
  externalDealId = NULL, flightEndTimeMs = NULL, flightStartTimeMs = NULL,
  inventoryDescription = NULL, isRfpTemplate = NULL,
  lastUpdateTimeMs = NULL, name = NULL, productId = NULL,
  productRevisionNumber = NULL, programmaticCreativeSource = NULL,
  proposalId = NULL, sellerContacts = NULL, sharedTargetings = NULL,
  syndicationProduct = NULL, terms = NULL, webPropertyCode = NULL)
}
\arguments{
\item{buyerPrivateData}{Buyer private data (hidden from seller)}

\item{creationTimeMs}{The time (ms since epoch) of the deal creation}

\item{creativePreApprovalPolicy}{Specifies the creative pre-approval policy (buyer-readonly)}

\item{creativeSafeFrameCompatibility}{Specifies whether the creative is safeFrame compatible (buyer-readonly)}

\item{dealId}{A unique deal-id for the deal (readonly)}

\item{dealServingMetadata}{Metadata about the serving status of this deal (readonly, writes via custom actions)}

\item{deliveryControl}{The set of fields around delivery control that are interesting for a buyer to see but are non-negotiable}

\item{externalDealId}{The external deal id assigned to this deal once the deal is finalized}

\item{flightEndTimeMs}{Proposed flight end time of the deal (ms since epoch) This will generally be stored in a granularity of a second}

\item{flightStartTimeMs}{Proposed flight start time of the deal (ms since epoch) This will generally be stored in a granularity of a second}

\item{inventoryDescription}{Description for the deal terms}

\item{isRfpTemplate}{Indicates whether the current deal is a RFP template}

\item{lastUpdateTimeMs}{The time (ms since epoch) when the deal was last updated}

\item{name}{The name of the deal}

\item{productId}{The product-id from which this deal was created}

\item{productRevisionNumber}{The revision number of the product that the deal was created from (readonly, except on create)}

\item{programmaticCreativeSource}{Specifies the creative source for programmatic deals, PUBLISHER means creative is provided by seller and ADVERTISR means creative is provided by buyer}

\item{proposalId}{No description}

\item{sellerContacts}{Optional Seller contact information for the deal (buyer-readonly)}

\item{sharedTargetings}{The shared targeting visible to buyers and sellers}

\item{syndicationProduct}{The syndication product associated with the deal}

\item{terms}{The negotiable terms of the deal}

\item{webPropertyCode}{No description}
}
\value{
MarketplaceDeal object
}
\description{
MarketplaceDeal Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A proposal can contain multiple deals. A deal contains the terms and targeting information that is used for serving.
}
