% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{MarketplaceDealParty}
\alias{MarketplaceDealParty}
\title{MarketplaceDealParty Object}
\usage{
MarketplaceDealParty(buyer = NULL, seller = NULL)
}
\arguments{
\item{buyer}{The buyer/seller associated with the deal}

\item{seller}{The buyer/seller associated with the deal}
}
\value{
MarketplaceDealParty object
}
\description{
MarketplaceDealParty Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

