% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_functions.R
\name{marketplacedeals.insert}
\alias{marketplacedeals.insert}
\title{Add new deals for the specified proposal}
\usage{
marketplacedeals.insert(AddOrderDealsRequest, proposalId)
}
\arguments{
\item{AddOrderDealsRequest}{The \link{AddOrderDealsRequest} object to pass to this method}

\item{proposalId}{proposalId for which deals need to be added}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adexchange.buyer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/ad-exchange/buyer-rest}{Google Documentation}

Other AddOrderDealsRequest functions: \code{\link{AddOrderDealsRequest}}
}
