% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{MarketplaceLabel}
\alias{MarketplaceLabel}
\title{MarketplaceLabel Object}
\usage{
MarketplaceLabel(accountId = NULL, createTimeMs = NULL,
  deprecatedMarketplaceDealParty = NULL, label = NULL)
}
\arguments{
\item{accountId}{The accountId of the party that created the label}

\item{createTimeMs}{The creation time (in ms since epoch) for the label}

\item{deprecatedMarketplaceDealParty}{Information about the party that created the label}

\item{label}{The label to use}
}
\value{
MarketplaceLabel object
}
\description{
MarketplaceLabel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

