% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{MarketplaceNote}
\alias{MarketplaceNote}
\title{MarketplaceNote Object}
\usage{
MarketplaceNote(creatorRole = NULL, dealId = NULL, note = NULL,
  noteId = NULL, proposalId = NULL, proposalRevisionNumber = NULL,
  timestampMs = NULL)
}
\arguments{
\item{creatorRole}{The role of the person (buyer/seller) creating the note}

\item{dealId}{Notes can optionally be associated with a deal}

\item{note}{The actual note to attach}

\item{noteId}{The unique id for the note}

\item{proposalId}{The proposalId that a note is attached to}

\item{proposalRevisionNumber}{If the note is associated with a proposal revision number, then store that here}

\item{timestampMs}{The timestamp (ms since epoch) that this note was created}
}
\value{
MarketplaceNote object
}
\description{
MarketplaceNote Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A proposal is associated with a bunch of notes which may optionally be associated with a deal and/or revision number.
}
