% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_functions.R
\name{marketplaceprivateauction.updateproposal}
\alias{marketplaceprivateauction.updateproposal}
\title{Update a given private auction proposal}
\usage{
marketplaceprivateauction.updateproposal(UpdatePrivateAuctionProposalRequest,
  privateAuctionId)
}
\arguments{
\item{UpdatePrivateAuctionProposalRequest}{The \link{UpdatePrivateAuctionProposalRequest} object to pass to this method}

\item{privateAuctionId}{The private auction id to be updated}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adexchange.buyer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/ad-exchange/buyer-rest}{Google Documentation}

Other UpdatePrivateAuctionProposalRequest functions: \code{\link{UpdatePrivateAuctionProposalRequest}}
}
