% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_functions.R
\name{performanceReport.list}
\alias{performanceReport.list}
\title{Retrieves the authenticated user's list of performance metrics.}
\usage{
performanceReport.list(accountId, endDateTime, startDateTime,
  maxResults = NULL, pageToken = NULL)
}
\arguments{
\item{accountId}{The account id to get the reports}

\item{endDateTime}{The end time of the report in ISO 8601 timestamp format using UTC}

\item{startDateTime}{The start time of the report in ISO 8601 timestamp format using UTC}

\item{maxResults}{Maximum number of entries returned on one result page}

\item{pageToken}{A continuation token, used to page through performance reports}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adexchange.buyer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/ad-exchange/buyer-rest}{Google Documentation}
}

