% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{PretargetingConfig.placements}
\alias{PretargetingConfig.placements}
\title{PretargetingConfig.placements Object}
\usage{
PretargetingConfig.placements()
}
\value{
PretargetingConfig.placements object
}
\description{
PretargetingConfig.placements Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other PretargetingConfig functions: \code{\link{PretargetingConfig.dimensions}},
  \code{\link{PretargetingConfig.excludedPlacements}},
  \code{\link{PretargetingConfig.videoPlayerSizes}},
  \code{\link{PretargetingConfig}},
  \code{\link{pretargetingConfig.insert}},
  \code{\link{pretargetingConfig.patch}},
  \code{\link{pretargetingConfig.update}}
}
