% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{PretargetingConfig}
\alias{PretargetingConfig}
\title{PretargetingConfig Object}
\usage{
PretargetingConfig(PretargetingConfig.dimensions = NULL,
  PretargetingConfig.excludedPlacements = NULL,
  PretargetingConfig.placements = NULL,
  PretargetingConfig.videoPlayerSizes = NULL, billingId = NULL,
  configId = NULL, configName = NULL, creativeType = NULL,
  dimensions = NULL, excludedContentLabels = NULL,
  excludedGeoCriteriaIds = NULL, excludedPlacements = NULL,
  excludedUserLists = NULL, excludedVerticals = NULL,
  geoCriteriaIds = NULL, isActive = NULL, languages = NULL,
  mobileCarriers = NULL, mobileDevices = NULL,
  mobileOperatingSystemVersions = NULL, placements = NULL,
  platforms = NULL, supportedCreativeAttributes = NULL, userLists = NULL,
  vendorTypes = NULL, verticals = NULL, videoPlayerSizes = NULL)
}
\arguments{
\item{PretargetingConfig.dimensions}{The \link{PretargetingConfig.dimensions} object or list of objects}

\item{PretargetingConfig.excludedPlacements}{The \link{PretargetingConfig.excludedPlacements} object or list of objects}

\item{PretargetingConfig.placements}{The \link{PretargetingConfig.placements} object or list of objects}

\item{PretargetingConfig.videoPlayerSizes}{The \link{PretargetingConfig.videoPlayerSizes} object or list of objects}

\item{billingId}{The id for billing purposes, provided for reference}

\item{configId}{The config id; generated automatically}

\item{configName}{The name of the config}

\item{creativeType}{List must contain exactly one of PRETARGETING_CREATIVE_TYPE_HTML or PRETARGETING_CREATIVE_TYPE_VIDEO}

\item{dimensions}{Requests which allow one of these (width, height) pairs will match}

\item{excludedContentLabels}{Requests with any of these content labels will not match}

\item{excludedGeoCriteriaIds}{Requests containing any of these geo criteria ids will not match}

\item{excludedPlacements}{Requests containing any of these placements will not match}

\item{excludedUserLists}{Requests containing any of these users list ids will not match}

\item{excludedVerticals}{Requests containing any of these vertical ids will not match}

\item{geoCriteriaIds}{Requests containing any of these geo criteria ids will match}

\item{isActive}{Whether this config is active}

\item{languages}{Request containing any of these language codes will match}

\item{mobileCarriers}{Requests containing any of these mobile carrier ids will match}

\item{mobileDevices}{Requests containing any of these mobile device ids will match}

\item{mobileOperatingSystemVersions}{Requests containing any of these mobile operating system version ids will match}

\item{placements}{Requests containing any of these placements will match}

\item{platforms}{Requests matching any of these platforms will match}

\item{supportedCreativeAttributes}{Creative attributes should be declared here if all creatives corresponding to this pretargeting configuration have that creative attribute}

\item{userLists}{Requests containing any of these user list ids will match}

\item{vendorTypes}{Requests that allow any of these vendor ids will match}

\item{verticals}{Requests containing any of these vertical ids will match}

\item{videoPlayerSizes}{Video requests satisfying any of these player size constraints will match}
}
\value{
PretargetingConfig object
}
\description{
PretargetingConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other PretargetingConfig functions: \code{\link{PretargetingConfig.dimensions}},
  \code{\link{PretargetingConfig.excludedPlacements}},
  \code{\link{PretargetingConfig.placements}},
  \code{\link{PretargetingConfig.videoPlayerSizes}},
  \code{\link{pretargetingConfig.insert}},
  \code{\link{pretargetingConfig.patch}},
  \code{\link{pretargetingConfig.update}}
}

