% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_functions.R
\name{pretargetingConfig.update}
\alias{pretargetingConfig.update}
\title{Updates an existing pretargeting config.}
\usage{
pretargetingConfig.update(PretargetingConfig, accountId, configId)
}
\arguments{
\item{PretargetingConfig}{The \link{PretargetingConfig} object to pass to this method}

\item{accountId}{The account id to update the pretargeting config for}

\item{configId}{The specific id of the configuration to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adexchange.buyer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/ad-exchange/buyer-rest}{Google Documentation}

Other PretargetingConfig functions: \code{\link{PretargetingConfig.dimensions}},
  \code{\link{PretargetingConfig.excludedPlacements}},
  \code{\link{PretargetingConfig.placements}},
  \code{\link{PretargetingConfig.videoPlayerSizes}},
  \code{\link{PretargetingConfig}},
  \code{\link{pretargetingConfig.insert}},
  \code{\link{pretargetingConfig.patch}}
}
