% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Price}
\alias{Price}
\title{Price Object}
\usage{
Price(amountMicros = NULL, currencyCode = NULL, expectedCpmMicros = NULL,
  pricingType = NULL)
}
\arguments{
\item{amountMicros}{The price value in micros}

\item{currencyCode}{The currency code for the price}

\item{expectedCpmMicros}{In case of CPD deals, the expected CPM in micros}

\item{pricingType}{The pricing type for the deal/product}
}
\value{
Price object
}
\description{
Price Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

