% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{PricePerBuyer}
\alias{PricePerBuyer}
\title{PricePerBuyer Object}
\usage{
PricePerBuyer(auctionTier = NULL, buyer = NULL, price = NULL)
}
\arguments{
\item{auctionTier}{Optional access type for this buyer}

\item{buyer}{The buyer who will pay this price}

\item{price}{The specified price}
}
\value{
PricePerBuyer object
}
\description{
PricePerBuyer Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Used to specify pricing rules for buyers/advertisers. Each PricePerBuyer in an product can become [0,1] deals. To check if there is a PricePerBuyer for a particular buyer or buyer/advertiser pair, we look for the most specific matching rule - we first look for a rule matching the buyer and advertiser, next a rule with the buyer but an empty advertiser list, and otherwise look for a matching rule where no buyer is set.
}

