% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{PrivateData}
\alias{PrivateData}
\title{PrivateData Object}
\usage{
PrivateData(referenceId = NULL, referencePayload = NULL)
}
\arguments{
\item{referenceId}{No description}

\item{referencePayload}{No description}
}
\value{
PrivateData object
}
\description{
PrivateData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

