% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Product}
\alias{Product}
\title{Product Object}
\usage{
Product(creationTimeMs = NULL, creatorContacts = NULL,
  deliveryControl = NULL, flightEndTimeMs = NULL,
  flightStartTimeMs = NULL, hasCreatorSignedOff = NULL,
  inventorySource = NULL, labels = NULL, lastUpdateTimeMs = NULL,
  legacyOfferId = NULL, marketplacePublisherProfileId = NULL, name = NULL,
  privateAuctionId = NULL, productId = NULL, publisherProfileId = NULL,
  publisherProvidedForecast = NULL, revisionNumber = NULL, seller = NULL,
  sharedTargetings = NULL, state = NULL, syndicationProduct = NULL,
  terms = NULL, webPropertyCode = NULL)
}
\arguments{
\item{creationTimeMs}{Creation time in ms}

\item{creatorContacts}{Optional contact information for the creator of this product}

\item{deliveryControl}{The set of fields around delivery control that are interesting for a buyer to see but are non-negotiable}

\item{flightEndTimeMs}{The proposed end time for the deal (ms since epoch) (buyer-readonly)}

\item{flightStartTimeMs}{Inventory availability dates}

\item{hasCreatorSignedOff}{If the creator has already signed off on the product, then the buyer can finalize the deal by accepting the product as is}

\item{inventorySource}{What exchange will provide this inventory (readonly, except on create)}

\item{labels}{Optional List of labels for the product (optional, buyer-readonly)}

\item{lastUpdateTimeMs}{Time of last update in ms}

\item{legacyOfferId}{Optional legacy offer id if this offer is a preferred deal offer}

\item{marketplacePublisherProfileId}{Marketplace publisher profile Id}

\item{name}{The name for this product as set by the seller}

\item{privateAuctionId}{Optional private auction id if this offer is a private auction offer}

\item{productId}{The unique id for the product (readonly)}

\item{publisherProfileId}{Id of the publisher profile for a given seller}

\item{publisherProvidedForecast}{Publisher self-provided forecast information}

\item{revisionNumber}{The revision number of the product}

\item{seller}{Information about the seller that created this product (readonly, except on create)}

\item{sharedTargetings}{Targeting that is shared between the buyer and the seller}

\item{state}{The state of the product}

\item{syndicationProduct}{The syndication product associated with the deal}

\item{terms}{The negotiable terms of the deal (buyer-readonly)}

\item{webPropertyCode}{The web property code for the seller}
}
\value{
Product object
}
\description{
Product Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A product is segment of inventory that a seller wishes to sell. It is associated with certain terms and targeting information which helps buyer know more about the inventory. Each field in a product can have one of the following setting:(readonly) - It is an error to try and set this field. (buyer-readonly) - Only the seller can set this field. (seller-readonly) - Only the buyer can set this field. (updatable) - The field is updatable at all times by either buyer or the seller.
}
