% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_functions.R
\name{products.get}
\alias{products.get}
\title{Gets the requested product by id.}
\usage{
products.get(productId)
}
\arguments{
\item{productId}{The id for the product to get the head revision for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adexchange.buyer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/ad-exchange/buyer-rest}{Google Documentation}
}

