% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Proposal}
\alias{Proposal}
\title{Proposal Object}
\usage{
Proposal(billedBuyer = NULL, buyer = NULL, buyerContacts = NULL,
  buyerPrivateData = NULL, dbmAdvertiserIds = NULL,
  hasBuyerSignedOff = NULL, hasSellerSignedOff = NULL,
  inventorySource = NULL, isRenegotiating = NULL, isSetupComplete = NULL,
  labels = NULL, lastUpdaterOrCommentorRole = NULL, name = NULL,
  negotiationId = NULL, originatorRole = NULL, privateAuctionId = NULL,
  proposalId = NULL, proposalState = NULL, revisionNumber = NULL,
  revisionTimeMs = NULL, seller = NULL, sellerContacts = NULL)
}
\arguments{
\item{billedBuyer}{Reference to the buyer that will get billed for this proposal}

\item{buyer}{Reference to the buyer on the proposal}

\item{buyerContacts}{Optional contact information of the buyer}

\item{buyerPrivateData}{Private data for buyer}

\item{dbmAdvertiserIds}{IDs of DBM advertisers permission to this proposal}

\item{hasBuyerSignedOff}{When an proposal is in an accepted state, indicates whether the buyer has signed off}

\item{hasSellerSignedOff}{When an proposal is in an accepted state, indicates whether the buyer has signed off Once both sides have signed off on a deal, the proposal can be finalized by the seller}

\item{inventorySource}{What exchange will provide this inventory (readonly, except on create)}

\item{isRenegotiating}{True if the proposal is being renegotiated (readonly)}

\item{isSetupComplete}{True, if the buyside inventory setup is complete for this proposal}

\item{labels}{List of labels associated with the proposal}

\item{lastUpdaterOrCommentorRole}{The role of the last user that either updated the proposal or left a comment}

\item{name}{The name for the proposal (updatable)}

\item{negotiationId}{Optional negotiation id if this proposal is a preferred deal proposal}

\item{originatorRole}{Indicates whether the buyer/seller created the proposal}

\item{privateAuctionId}{Optional private auction id if this proposal is a private auction proposal}

\item{proposalId}{The unique id of the proposal}

\item{proposalState}{The current state of the proposal}

\item{revisionNumber}{The revision number for the proposal (readonly)}

\item{revisionTimeMs}{The time (ms since epoch) when the proposal was last revised (readonly)}

\item{seller}{Reference to the seller on the proposal}

\item{sellerContacts}{Optional contact information of the seller (buyer-readonly)}
}
\value{
Proposal object
}
\description{
Proposal Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a proposal in the marketplace. A proposal is the unit of negotiation between a seller and a buyer and contains deals which are served. Each field in a proposal can have one of the following setting:(readonly) - It is an error to try and set this field. (buyer-readonly) - Only the seller can set this field. (seller-readonly) - Only the buyer can set this field. (updatable) - The field is updatable at all times by either buyer or the seller.
}
\seealso{
Other Proposal functions: \code{\link{proposals.patch}},
  \code{\link{proposals.update}}
}
