% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_functions.R
\name{proposals.patch}
\alias{proposals.patch}
\title{Update the given proposal. This method supports patch semantics.}
\usage{
proposals.patch(Proposal, proposalId, revisionNumber, updateAction)
}
\arguments{
\item{Proposal}{The \link{Proposal} object to pass to this method}

\item{proposalId}{The proposal id to update}

\item{revisionNumber}{The last known revision number to update}

\item{updateAction}{The proposed action to take on the proposal}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adexchange.buyer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/ad-exchange/buyer-rest}{Google Documentation}

Other Proposal functions: \code{\link{Proposal}},
  \code{\link{proposals.update}}
}
