% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{PublisherProfileApiProto}
\alias{PublisherProfileApiProto}
\title{PublisherProfileApiProto Object}
\usage{
PublisherProfileApiProto(accountId = NULL, audience = NULL,
  buyerPitchStatement = NULL, directContact = NULL, exchange = NULL,
  googlePlusLink = NULL, isParent = NULL, isPublished = NULL,
  logoUrl = NULL, mediaKitLink = NULL, name = NULL, overview = NULL,
  profileId = NULL, programmaticContact = NULL, publisherDomains = NULL,
  publisherProfileId = NULL, publisherProvidedForecast = NULL,
  rateCardInfoLink = NULL, samplePageLink = NULL, seller = NULL,
  state = NULL, topHeadlines = NULL)
}
\arguments{
\item{accountId}{The account id of the seller}

\item{audience}{Publisher provided info on its audience}

\item{buyerPitchStatement}{A pitch statement for the buyer}

\item{directContact}{Direct contact for the publisher profile}

\item{exchange}{Exchange where this publisher profile is from}

\item{googlePlusLink}{Link to publisher's Google+ page}

\item{isParent}{True, if this is the parent profile, which represents all domains owned by the publisher}

\item{isPublished}{True, if this profile is published}

\item{logoUrl}{The url to the logo for the publisher}

\item{mediaKitLink}{The url for additional marketing and sales materials}

\item{name}{No description}

\item{overview}{Publisher provided overview}

\item{profileId}{The pair of (seller}

\item{programmaticContact}{Programmatic contact for the publisher profile}

\item{publisherDomains}{The list of domains represented in this publisher profile}

\item{publisherProfileId}{Unique Id for publisher profile}

\item{publisherProvidedForecast}{Publisher provided forecasting information}

\item{rateCardInfoLink}{Link to publisher rate card}

\item{samplePageLink}{Link for a sample content page}

\item{seller}{Seller of the publisher profile}

\item{state}{State of the publisher profile}

\item{topHeadlines}{Publisher provided key metrics and rankings}
}
\value{
PublisherProfileApiProto object
}
\description{
PublisherProfileApiProto Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
