% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{PublisherProvidedForecast}
\alias{PublisherProvidedForecast}
\title{PublisherProvidedForecast Object}
\usage{
PublisherProvidedForecast(dimensions = NULL, weeklyImpressions = NULL,
  weeklyUniques = NULL)
}
\arguments{
\item{dimensions}{Publisher provided dimensions}

\item{weeklyImpressions}{Publisher provided weekly impressions}

\item{weeklyUniques}{Publisher provided weekly uniques}
}
\value{
PublisherProvidedForecast object
}
\description{
PublisherProvidedForecast Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This message carries publisher provided forecasting information.
}
