% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Seller}
\alias{Seller}
\title{Seller Object}
\usage{
Seller(accountId = NULL, subAccountId = NULL)
}
\arguments{
\item{accountId}{The unique id for the seller}

\item{subAccountId}{Optional sub-account id for the seller}
}
\value{
Seller object
}
\description{
Seller Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
