% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{SharedTargeting}
\alias{SharedTargeting}
\title{SharedTargeting Object}
\usage{
SharedTargeting(exclusions = NULL, inclusions = NULL, key = NULL)
}
\arguments{
\item{exclusions}{The list of values to exclude from targeting}

\item{inclusions}{The list of value to include as part of the targeting}

\item{key}{The key representing the shared targeting criterion}
}
\value{
SharedTargeting object
}
\description{
SharedTargeting Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

