% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{TargetingValue}
\alias{TargetingValue}
\title{TargetingValue Object}
\usage{
TargetingValue(creativeSizeValue = NULL, dayPartTargetingValue = NULL,
  longValue = NULL, stringValue = NULL)
}
\arguments{
\item{creativeSizeValue}{The creative size value to exclude/include}

\item{dayPartTargetingValue}{The daypart targeting to include / exclude}

\item{longValue}{The long value to exclude/include}

\item{stringValue}{The string value to exclude/include}
}
\value{
TargetingValue object
}
\description{
TargetingValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

