% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{TargetingValueCreativeSize}
\alias{TargetingValueCreativeSize}
\title{TargetingValueCreativeSize Object}
\usage{
TargetingValueCreativeSize(companionSizes = NULL, creativeSizeType = NULL,
  size = NULL, skippableAdType = NULL)
}
\arguments{
\item{companionSizes}{For video size type, the list of companion sizes}

\item{creativeSizeType}{The Creative size type}

\item{size}{For regular or video creative size type, specifies the size of the creative}

\item{skippableAdType}{The skippable ad type for video size}
}
\value{
TargetingValueCreativeSize object
}
\description{
TargetingValueCreativeSize Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
