% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{TargetingValueDayPartTargeting}
\alias{TargetingValueDayPartTargeting}
\title{TargetingValueDayPartTargeting Object}
\usage{
TargetingValueDayPartTargeting(dayParts = NULL, timeZoneType = NULL)
}
\arguments{
\item{dayParts}{No description}

\item{timeZoneType}{No description}
}
\value{
TargetingValueDayPartTargeting object
}
\description{
TargetingValueDayPartTargeting Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
