% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{TargetingValueDayPartTargetingDayPart}
\alias{TargetingValueDayPartTargetingDayPart}
\title{TargetingValueDayPartTargetingDayPart Object}
\usage{
TargetingValueDayPartTargetingDayPart(dayOfWeek = NULL, endHour = NULL,
  endMinute = NULL, startHour = NULL, startMinute = NULL)
}
\arguments{
\item{dayOfWeek}{No description}

\item{endHour}{No description}

\item{endMinute}{No description}

\item{startHour}{No description}

\item{startMinute}{No description}
}
\value{
TargetingValueDayPartTargetingDayPart object
}
\description{
TargetingValueDayPartTargetingDayPart Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

