% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{TargetingValueSize}
\alias{TargetingValueSize}
\title{TargetingValueSize Object}
\usage{
TargetingValueSize(height = NULL, width = NULL)
}
\arguments{
\item{height}{The height of the creative}

\item{width}{The width of the creative}
}
\value{
TargetingValueSize object
}
\description{
TargetingValueSize Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

