% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{UpdatePrivateAuctionProposalRequest}
\alias{UpdatePrivateAuctionProposalRequest}
\title{UpdatePrivateAuctionProposalRequest Object}
\usage{
UpdatePrivateAuctionProposalRequest(externalDealId = NULL, note = NULL,
  proposalRevisionNumber = NULL, updateAction = NULL)
}
\arguments{
\item{externalDealId}{The externalDealId of the deal to be updated}

\item{note}{Optional note to be added}

\item{proposalRevisionNumber}{The current revision number of the proposal to be updated}

\item{updateAction}{The proposed action on the private auction proposal}
}
\value{
UpdatePrivateAuctionProposalRequest object
}
\description{
UpdatePrivateAuctionProposalRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other UpdatePrivateAuctionProposalRequest functions: \code{\link{marketplaceprivateauction.updateproposal}}
}
